adk/requirements.txt
google-adk
yfinance==0.2.56
psutil==5.9.5
litellm
google-generativeai
python-dotenv